requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
scikit-learn==1.3.0
faiss-cpu==1.7.4
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3
tiktoken==0.5.1 